requirements.txt
dcor==0.6
h5py==3.8.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
Pillow==10.1.0
scikit_learn==1.3.2
scipy==1.11.4
torch==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.64.1
